updated Event model
change bookedTickets from long to int

+++ src/main/java/com/eventtickets/datatier/controllers/DTO/EventDTO.java
@@ -22,5 +22,5 @@ public class EventDTO {
 	private LocalDateTime timeOfTheEvent;
 	private double ticketPrice;
 	private long organizerId;
-	private long bookedTickets;
+	private int bookedTickets;
 }
